backend/requirements.txt
Flask
flask-cors
python-dotenv
PyMuPDF
openai
google-genai
requests
